switched to https://github.com/Siddhant-Naik/TheFlash-Kernel-A5-A7-2017 kernel tree

+++ drivers/clk/samsung/composite.c
@@ -53,7 +53,7 @@ struct samsung_clk_reg_dump *samsung_clk_alloc_reg_dump(
 						unsigned long nr_rdump)
 {
 	struct samsung_clk_reg_dump *rd;
-	unsigned long i;
+	unsigned int i;
 
 	rd = kcalloc(nr_rdump, sizeof(*rd), GFP_KERNEL);
 	if (!rd)
